ConstructAppThreeComponents
